policy for discounts

+++ modules/Abd/Discount/src/Http/Controllers/DiscountController.php
@@ -14,6 +14,7 @@ class DiscountController extends Controller
 {
     public function index(CourseRepo $courseRepo, DiscountRepo $discountRepo)
     {
+        $this->authorize("manage", Discount::class);
         $discounts = $discountRepo->paginateAll();
         $courses = $courseRepo->getAll(Course::CONFIRMATION_STATUS_ACCEPTED);
         return view('Discounts::index', compact('courses', 'discounts'));
@@ -21,6 +22,7 @@ class DiscountController extends Controller
 
     public function store(DiscountRequest $request, DiscountRepo $repo)
     {
+        $this->authorize("manage", Discount::class);
         $repo->store($request->all());
         newFeedback();
         return back();
@@ -28,12 +30,14 @@ class DiscountController extends Controller
 
     public function edit(Discount $discount,CourseRepo $courseRepo)
     {
+        $this->authorize("manage", Discount::class);
         $courses = $courseRepo->getAll(Course::CONFIRMATION_STATUS_ACCEPTED);
         return view('Discounts::edit', compact('discount', 'courses'));
     }
 
     public function update(Discount $discount, DiscountRequest $request, DiscountRepo $repo)
     {
+        $this->authorize("manage", Discount::class);
         $repo->update($discount->id, $request->all());
         newFeedback();
         return redirect()->route("discounts.index");
@@ -41,6 +45,7 @@ class DiscountController extends Controller
 
     public function destroy(Discount $discount)
     {
+        $this->authorize("manage", Discount::class);
         $discount->delete();
         return AjaxResponses::SuccessResponse();
     }
